Move pre-selection test for belongs to case to end-to-end spec
Kept getting error of the form

    undefined method `admin_product_product_variant_path' for
    #<Admin::VariantsController>

when trying execute the test as part of the request spec. When running
the test individually everything worked, but once the end-to-end test
had run, the error occurred. I therefore suspect this might be caused
by setting up the nested Active Admin resources twice.

Testing both cases as part of the end-to-end spec helps work around
the problem.